requirements.txt
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
configparser==5.0.1
constantly==15.1.0
crayons==0.4.0
cryptography==3.3.1
cssselect==1.1.0
fake-useragent==0.1.11
filelock==3.0.12
geckodriver-autoinstaller==0.1.0
google-api-core==1.25.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
jmespath==0.10.0
lxml==4.6.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
parsel==1.6.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-file==1.5.1
rsa==4.7
selenium==3.141.0
selenium-firefox==0.5.34
six==1.15.0
soupsieve==2.1
tldextract==3.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
w3lib==1.22.0
webdriver-manager==3.3.0
Werkzeug==1.0.1
zope.interface==5.2.0
